Cocoon Cleanup & Minor Bloodsucker Tweaks (#1058)
# Description

- Generalizes cocooning
- Allows any mob to be cocooned
- Cocoon bloodsucking moved to vampirism system
- Any blood sucker can drink from cocoons
- Vampirism no longer fails if bloodstream isn't normal blood, but gives
a pop up
- Vampirism `WebRequired` actually works in a way that makes sense
- Adds cocooning and bloodsucker to all spider mobs + Arachnids

resolves #978

---

# Changelog

:cl:
- tweak: All spiders, arachne, and arachnids can cocoon mobs, and drink
their blood.

---------
