+++ Content.Client/Cocoon/CocoonSystem.cs
@@ -1,4 +1,4 @@
-using Content.Shared.Arachne;
+using Content.Shared.Cocoon;
 using Content.Shared.Humanoid;
 using Robust.Client.GameObjects;
 using Robust.Shared.Containers;
