--- Content.Shared/Arachne/Events.cs
@@ -1,11 +1,0 @@
-using Robust.Shared.Map;
-using Robust.Shared.Serialization;
-using Content.Shared.DoAfter;
-
-namespace Content.Shared.Arachne
-{
-    [Serializable, NetSerializable]
-    public sealed partial class ArachneCocoonDoAfterEvent : SimpleDoAfterEvent
-    {
-    }
-}
