--- Content.Shared/Arachne/WebComponent.cs
@@ -1,8 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared.Arachne
-{
-    [RegisterComponent, NetworkedComponent]
-    public sealed partial class WebComponent : Component
-    {}
-}
