+++ Content.Shared/Cocoon/CocoonDoAfterEvent.cs
@@ -1,0 +1,10 @@
+using Robust.Shared.Serialization;
+using Content.Shared.DoAfter;
+
+namespace Content.Shared.Cocoon
+{
+    [Serializable, NetSerializable]
+    public sealed partial class CocoonDoAfterEvent : SimpleDoAfterEvent
+    {
+    }
+}
